Implemented Scorekeeping

+++ Weapon/Entity/entities.hpp
@@ -46,12 +46,12 @@ class playerData {
 //Hence it actually is another layer of abstraction; ApplicationLogic.
 class weaponSettings {
 	private:
-		std::array<const char *, 5> weaponNames = 	{	"Shotgun", "Pistol", 	"Sniper", 	"AK-47", 	"M16"	};
-		std::array<int, 5> damageFactors = 			{		10,		15, 		40, 		6, 			10		};
-		std::array<int, 5> bulletsPerMag = 			{		6,		17, 		5, 			30, 		30		};
-		std::array<int, 5> maxShotsPerTenSeconds = 	{		40, 	100, 		20,			100, 		140		};
-		std::array<bool, 5> autoAllowed = 			{		false, 	false, 		false, 		true, 		true	};
-		std::array<bool, 5> burstAllowed = 			{		false, 	false, 		false, 		true, 		true	};
+		std::array<const char *, 5> weaponNames = 		{	"Shotgun", "Pistol", 	"Sniper", 	"AK-47", 	"M16"	};
+		std::array<uint8_t, 5> damageFactors = 			{		10,		15, 		40, 		6, 			10		};
+		std::array<uint8_t, 5> bulletsPerMag = 			{		6,		17, 		5, 			30, 		30		};
+		std::array<uint8_t, 5> maxShotsPerTenSeconds = 	{		40, 	100, 		20,			100, 		140		};
+		std::array<bool, 5> autoAllowed = 				{		false, 	false, 		false, 		true, 		true	};
+		std::array<bool, 5> burstAllowed = 				{		false, 	false, 		false, 		true, 		true	};
 	public:
 		const char * getName(const int weaponId) const;
 		const int getDamage(const int weaponId, const int distance) const;
